requirements.txt
python-dotenv==1.1.1
pydantic==2.11.7
